+++ index.html
@@ -164,6 +164,8 @@
       </div>
     </section>
   </template>
+  <script src="js/data.js" type="module"></script>
+  <script src="js/util.js" type="module"></script>
   <script src="js/main.js"></script>
 </body>
 </html>
